pyproject.toml
[tool.poetry]
name = "ai-builders-24"
version = "0.1.0"
description = ""
authors = ["vintro <vince@plasticlabs.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
honcho-ai = "^0.0.12"
mirascope = {extras = ["anthropic"], version = "^0.18.1"}
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
